cleanups: 1. moved LineBuffer to inner class of the Export Mapper 2. extended support to 9.x direct copy 3. Added test case
